src/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
SQLAlchemy-Utils==0.33.1
Flask-WTF==0.14.2
ccxt==1.17.169
APScheduler==3.5.1
pytest==3.8.1
requests==2.19.1
arrow==0.12.1
pytest-mock==1.10.0
pyinstaller==3.4
